+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## [0.6.4] - 2026-08-25
+
+### Added
+
+- Make the Uncommitted Changes graph row selectable and show its modified, added, deleted, renamed, and untracked files in the details pane
+- Open HEAD-to-working-tree diffs directly from the uncommitted file list
+
+### Changed
+
+- Refresh an open uncommitted file list automatically when repository files change
+
 ## [0.6.3] - 2026-08-25
 
 ### Added
@@ -104,7 +115,8 @@
 
 Initial release
 
-[Unreleased]: https://github.com/gerhardtEyes/zeeho-git-graph/compare/v0.6.3...HEAD
+[Unreleased]: https://github.com/gerhardtEyes/zeeho-git-graph/compare/v0.6.4...HEAD
+[0.6.4]: https://github.com/gerhardtEyes/zeeho-git-graph/compare/v0.6.3...v0.6.4
 [0.6.3]: https://github.com/gerhardtEyes/zeeho-git-graph/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/gerhardtEyes/zeeho-git-graph/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/gerhardtEyes/zeeho-git-graph/releases/tag/v0.6.1